refactor(architecture):  modules auto provide
