+++ packages/framework/src/base/middleware.ts
@@ -6,10 +6,10 @@
  */
 import type { TNext } from '../middleware';
 import type { Request } from '../request';
-import { ComponentType } from '../symbol';
 import { Base } from './base';
+import { componentType } from '../decorators/component-type';
 
-@Reflect.metadata('type', ComponentType.Middleware)
+@componentType('middleware')
 export abstract class Middleware extends Base {
   abstract resolve(request: Request, next: TNext): any | Promise<any>
 }
