+++ packages/framework/src/component/index.ts
@@ -1,0 +1,1 @@
+export * from './component-provider';
