+++ packages/framework/src/foundation/auto-providers/index.ts
@@ -1,4 +1,4 @@
-export * from './built-in';
+export * from './app';
 export * from './common';
 export * from './http-server';
-export * from './resolver';
+
