+++ packages/framework/src/resource/index.ts
@@ -7,3 +7,4 @@
 
 export * from './item';
 export * from './collection';
+export * from './resource-provider';
